Added files via upload
This program produces a cleveryly constructed brotli file that decompresses fine, but the compressor cannot compress it back as small as it was. It is written as inspiration for the the brotli compressor artists. The file format is a format that I actually used for something: a file with lines of random hex data with simple formatting.
I got the idea when writing brotlidump.py after some experimentation.